% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chr_axis_creator}
\alias{chr_axis_creator}
\title{Function that generates the x axis ticks for chromosome positions}
\usage{
chr_axis_creator(geneinfo.df)
}
\arguments{
\item{geneinfo.df}{dataframe that holds the position of the genes.}
}
\description{
The function takes in a gene position info data frame with the following columns:
"phenotypes" holding the names of the genes used in the analysis (rownames of phenotype.mx)
"pheno_chr" holding the chromosome each gene belongs to
"pheno_start" the starting position of a given gene
"pheno_end" the end position of a given gene
"idx" a non-overlapping index that shows orders the genes based on their position, which goes from
1 to number of genes.
}

