% Generated by roxygen2 (4.0.2): do not edit by hand
\name{covar_association_check}
\alias{covar_association_check}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
covar_association_check(input_list = NULL, covars = NULL,
  cor_threshold = 0.05)
}
\arguments{
\item{input_list}{An input ICAobject or PCAobject.}

\item{covars}{A matrix of covariates with each row for a sample and each column for a covaraite}

\item{cor_threshold}{Threshold that is going to be used for identifying significant associations.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

