% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateAssociationCheck.R
\name{covarAssociationCheck}
\alias{covarAssociationCheck}
\title{Calculate association between components and given covariates}
\usage{
covarAssociationCheck(input_list = NULL, covars = NULL, col_names = NULL,
  cor_threshold = 0.05)
}
\arguments{
\item{input_list}{ICA or PCA object generated by \code{runICA()} or
\code{runPCA()}.}

\item{covars}{A dataframe of covariates with each row for a sample
and each column for a covariate.}

\item{col_names}{Column names of \code{covars} to
be used for association testing. Default is to test every input covariate.}

\item{cor_threshold}{Bonferroni threshold that is going to be used for
identifying significant associations.}
}
\value{
The output will be the \code{input_list} with additional entries
        \code{comp_cov, covars, covar_threshold, covar_pvals}.
        \code{comp_cov} will contain a list with entries for each component
                        showing the associated covariate and p-value.
        \code{covars} will contain a copy of sample_info to keep all
                      the data in one place.
        \code{covar_threshold} will record the threshold
                               used to call significant associations.
        \code{covar_pvals} will contain the matrix of p-values
                           for all covariate and component pairs.
}
\description{
Check the association between sample covariates and
independent/principal components.
}
\examples{
data(expr_data, sample_info)

pca_result <- runPCA(expr_data)

pca_result <- covarAssociationCheck(pca_result, covars = sample_info)
}
