% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clusters.R
\name{detect_clusters}
\alias{detect_clusters}
\title{Function to estimate the number of clusters in IC coefficients or PC projections.}
\usage{
detect_clusters(input_list = NULL)
}
\arguments{
\item{input_list}{An input ICAobject or PCAobject.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

