% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectClusters.R
\name{detectClusters}
\alias{detectClusters}
\title{Function to estimate the number of clusters in IC coefficients or PC projections.}
\usage{
detectClusters(input_list = NULL)
}
\arguments{
\item{input_list}{input_list ICA or PCA object generated by \code{run_ica()} or \code{runPCA()}.}
}
\value{
Running \code{detectClusters()} will generate a copy of the input_list with additional entries \code{n_clust,mclust_result}.
        \code{n_clust} will contain the number of estimated clusters for each component, and \code{mclust_result} is a list with
        each result of \code{mclust} saved for all components.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\examples{

data(expr_data)

pca_result <- runPCA(expr_data)
pca_result <- detectClusters(pca_result)

}
