% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fastICA_gene_expr}
\alias{fastICA_gene_expr}
\title{ICA analysis based on the package fastICA}
\usage{
fastICA_gene_expr(X, n.comp, fun = "logcosh", alpha = 1,
  scale_pheno = FALSE, maxit = 200, tol = 1e-04, verbose = TRUE,
  w.init = NULL)
}
\arguments{
\item{X}{Phenotype matrix with diemnsions g x N}

\item{n.comp}{Number of components to be estimated or method to estimate it.}

\item{fun}{Function to be used in ICA estimation}

\item{alpha}{alpha for ICA, should be in range [1,2].}

\item{scale_pheno}{Logical value specifying the scaling of rows of X.}

\item{maxit}{Maximum iterations}

\item{tol}{Threshold for convergence}

\item{verbose}{If TRUE details of the estimation process are shown.}

\item{w.init}{Initial value for W, if left unspecified random numbers will be used.}
}
\value{
List with the following entries.
}
\description{
Minor tweak to fastICA to make it faster
}
\keyword{keywords}

