% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate_mx.R
\name{get_covariate_mx}
\alias{get_covariate_mx}
\title{Function that returns a matrix of IC coefficients that can be used as covariates.}
\usage{
get_covariate_mx(input_list = NULL, idx = NULL)
}
\arguments{
\item{input_list}{An input ICAobject or PCAobject.}

\item{idx}{Component indexes that the user wishes to recover.}
}
\value{
Matrix with component coefficients
}
\description{
Returning a matrix of IC coefficients that are associated with covariates or have multiple clusters.
}
\keyword{keywords}

