% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Retrieve information for top contributing genes for a given IC or PC}
\usage{
get_gene_info(input_list = NULL, comp_idx = NULL, geneinfo_df = NULL,
  n_peaks = 10)
}
\arguments{
\item{input_list}{ICA or PCA object generated by `run_ica()` or `run_pca()`}

\item{comp_idx}{Component number to be plotted.}

\item{geneinfo_df}{Dataframe containing information about the gene positions. This is used to generate the gene weight plot}

\item{n_peaks}{Number of top contributors to show. Default is set to 10.}
}
\value{
dataframe with positional information of genes
}
\description{
Generate a dataframe with gene information for top contributors for a given IC or PC
}
\examples{
data(expr_data, probe_info)

ica_object <- run_ica(expr_data)
get_gene_info(input_list = ica_object,
              comp_idx = 1,
              geneinfo_df = probe_info,
              n_peaks = 10)

}

