% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovariateMx.R
\name{getCovariateMx}
\alias{getCovariateMx}
\title{Function that returns a matrix of IC coefficients that can be used as covariates.}
\usage{
getCovariateMx(input_list = NULL, idx = NULL)
}
\arguments{
\item{input_list}{ICA or PCA object generated by \code{runICA()} or \code{runPCA()}.}

\item{idx}{Component indexes that the user wishes to recover.}
}
\value{
Matrix with component coefficients
}
\description{
Returning a matrix of IC coefficients that are associated with covariates or have multiple clusters.
}
\examples{
data(expr_data)

ica_object <- runICA(expr_data)
ica_object <- covarAssociationCheck(ica_object, covars = sample_info)
ica_object <- detectClusters(ica_object)
getCovariateMx(ica_object)

}
\keyword{keywords}
