% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Retrieve information for top contributing genes for a given IC or PC}
\usage{
getGeneInfo(input_list = NULL, comp_idx = NULL, geneinfo_df = NULL,
  n_peaks = 10)
}
\arguments{
\item{input_list}{ICA or PCA object generated by `runICA()` or `runPCA`}

\item{comp_idx}{Component number to be plotted.}

\item{geneinfo_df}{Dataframe containing information about the gene positions.
This is used to generate the gene weight plot.}

\item{n_peaks}{Number of top contributors to show. Default is set to 10.}
}
\value{
dataframe with positional information of genes
}
\description{
Generate a dataframe with gene information for
top contributors for a given IC or PC
}
\examples{
data(expr_data, probe_info)

ica_result <- runICA(expr_data)
getGeneInfo(input_list = ica_result,
              comp_idx = 1,
              geneinfo_df = probe_info,
              n_peaks = 10)

}
