% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_association_check.R
\name{ic_covariate_association_test}
\alias{ic_covariate_association_test}
\title{Testing the association between independent component coefficients and known covariates.}
\usage{
ic_covariate_association_test(input.A, info.input)
}
\arguments{
\item{input.A}{A matrix of an ica_result list}

\item{info.input}{A dataframe that holds the measured covariates for each sample}
}
\value{
output A matrix holding the p-values for each indepenedent component and covariate pair.
}
\description{
The function takes in the A matrix from the ica.result object
with a dataframe of measured covariates to test the association
between them.
}
\keyword{keywords}

