% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IC_variance_calc}
\alias{IC_variance_calc}
\title{Calculating the variation explained by each component}
\usage{
IC_variance_calc(s, a)
}
\arguments{
\item{s}{A single column of the S matrix with dimensions 1 x g  \code{s}}

\item{a}{A single row of the A matrix with dimensions N x 1 \code{a}}
}
\value{
A single number of how much variantion is explained by a single IC.
}
\description{
The function takes in a single row of the ICA result A matrix and a
single column of the ICA result S matrix and calculates the sums of squares.
}

