% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastICA_gene_expr.R
\name{ica_R_par}
\alias{ica_R_par}
\title{ICA parameter estimation function}
\usage{
ica_R_par(X, n.comp, tol, fun, alpha, maxit, verbose, w.init)
}
\arguments{
\item{X}{Phenotype matrix with diemnsions g x N}

\item{n.comp}{Number of components to be estimated or method to estimate it.}

\item{tol}{Threshold for convergence}

\item{fun}{Function to be used in ICA estimation}

\item{alpha}{alpha for ICA, should be in range [1,2].}

\item{maxit}{Maximum iterations}

\item{verbose}{If TRUE details of the estimation process are shown.}

\item{w.init}{Initial value for W, if left unspecified random numbers will be used.}
}
\value{
List with the following entries.
}
\description{
Adopted from fastICA package
}
\keyword{keywords}

