% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peak_detection}
\alias{peak_detection}
\title{Labeling peaks for each independent component}
\usage{
peak_detection(ica.input.s)
}
\arguments{
\item{ica.input.s}{A single column of the S matrix with dimensions 1 x g  \code{s}}
}
\value{
A vector that contains the gene weights of the signficant peaks for a single independent component sorted
in the decreasing order of absolute magnitude.
}
\description{
The function takes in the A matrix from the ica.result object
with a dataframe of measured covariates to test the association
between them.
}
\keyword{keywords}

