% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_component}
\alias{plot_component}
\title{Generate plots for single component}
\usage{
plot_component(input_list = NULL, comp_idx, geneinfo_df = NULL,
  plot_here = TRUE)
}
\arguments{
\item{input_list}{ICA or PCA object generated by `run_ica()` or `run_pca()`}

\item{comp_idx}{Component number to be plotted.}

\item{geneinfo_df}{Dataframe containing information about the gene positions. This is used to generate the gene weight plot}

\item{plot_here}{If set to TRUE it will generate the plots in the current session, and if set to FALSE it will return a list of plots only
and not generate plots in the current session.}
}
\value{
List of plots
}
\description{
Generate gene weight and coefficient plots for a single independent or principal component.
}
\examples{
data(expr_data)

pca_object <- run_pca(expr_data)
plot_component(pca_object, comp_idx = 1)

}

