% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_info}
\alias{probe_info}
\title{Positional information for 2642 probes used in \code{expr_data}}
\format{A matrix with 2642 rows and 4 columns:
\describe{
  \item{phenotype}{The probe id used in \code{expr_data}}
  \item{pheno_chr}{Chromosome of measured probe.}
  \item{pheno_start}{Starting position of measurement probe}
  \item{pheno_end}{End position of measurement probe}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60028}
}
\usage{
probe_info
}
\description{
The original dataset is described in the following paper:
Dhingra N, Shemer A, Correa da Rosa J, Rozenblit M et al.
Molecular profiling of contact dermatitis skin identifies allergen-dependent differences in immune response.
J Allergy Clin Immunol 2014 Aug;134(2):362-72. PMID: 24768652
}
\keyword{datasets}
