% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ica.R
\name{run_ica}
\alias{run_ica}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
run_ica(pheno_mx = NULL, k_est = NULL, var_cutoff = 99, n_runs = 1,
  n_cores = NULL, scale_pheno = FALSE, h_clust_cutoff = 0.3,
  max_iter = 10, similarity_measure = "peaks")
}
\arguments{
\item{pheno_mx}{Phenotype matrix with diemnsions g x N, where g is the number of genes and N is the number of samples.}

\item{k_est}{Number of components to be estimated or method to estimate it.}

\item{var_cutoff}{Percent variance threshold to use when <k_est> is not supplied.}

\item{n_runs}{Number of times to run ICA. If this is set to a number larger than 1,
only ICs that replicate between runs are going to be returned}

\item{n_cores}{Number of cores to use when n_runs is larger than 1.}

\item{scale_pheno}{If set to TRUE the pre-processing step of the data will include a scaling step.
This will divide all phenotypes by their standard deviation.}

\item{h_clust_cutoff}{is the cutoff value used in hierarchical clustering. Default is set to 0.3.}

\item{max_iter}{Maximum iterations for estimating k for each run. Default value is set to 10.}

\item{similarity_measure}{How to measure the similarity between ICs.
If set to "peaks" only gene weights that are greater than 1 sd are used to calculate similarity.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

