% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Custom PCA function for analyzing gene expression data.}
\usage{
run_pca(pheno_mx = NULL, scale_pheno = FALSE)
}
\arguments{
\item{pheno_mx}{Phenotype matrix with diemnsions g x N, where g is the number of genes and N is the number of samples.}

\item{scale_pheno}{Logical value specifying the scaling of row of the pheno_mx. Default is set to FALSE.}
}
\value{
List with the following entries
}
\description{
Performing PCA on a dataset and create a list object with results.
}
\keyword{keywords}

