% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runICA.R
\name{runICA}
\alias{runICA}
\title{Custom ICA function for analyzing gene expression data.}
\usage{
runICA(pheno_mx = NULL, se_obj = NULL, assay_idx = NULL, k_est = NULL,
  var_cutoff = 99, n_runs = 1, n_cores = NULL, scale_pheno = FALSE,
  h_clust_cutoff = 0.3, max_iter = 10, similarity_measure = "peaks")
}
\arguments{
\item{pheno_mx}{Phenotype matrix with diemnsions g x N, where g is the number of genes and N is the number of samples.}

\item{se_obj}{The input for the function can also be a \code{SummarizedExperiment} or \code{RangedSummarizedExperiment} object created through the
\code{SummarizedExperiment} package on bioconductor.}

\item{assay_idx}{The assay index to be used to retrieved a single assay from the SummarizedExperiment object.}

\item{k_est}{Number of components to be estimated or method to estimate it.}

\item{var_cutoff}{Percent variance threshold to use when <k_est> is not supplied.}

\item{n_runs}{Number of times to run ICA. If this is set to a number larger than 1,
only ICs that replicate between runs are going to be returned}

\item{n_cores}{Number of cores to use when n_runs is larger than 1.}

\item{scale_pheno}{If set to TRUE the pre-processing step of the data will include a scaling step.
This will divide all phenotypes by their standard deviation.}

\item{h_clust_cutoff}{is the cutoff value used in hierarchical clustering. Default is set to 0.3.}

\item{max_iter}{Maximum iterations for estimating k for each run. Default value is set to 10.}

\item{similarity_measure}{How to measure the similarity between ICs.
If set to "peaks" only gene weights that are greater than 1 sd are used to calculate similarity.}
}
\value{
The following entries will be generated in the output list \code{ica_object} after running the example above. \cr
\code{A} : The IC coefficient matrix, with each row representing coefficients for the corresponding independent component. (standard \code{fastICA()} output) \cr
\code{S} : Matrix of gene weights for each independent component. Each column holds a single component. (standard \code{fastICA()} output)\cr
\code{percent_var} : The percent variance each independent component is explaining.\cr
\code{peaks} : Indicating which gene has a gene weight larger than 2 standard deviations of its component gene weights.\cr
\code{order} : The order of independent components based on the variance that they explain.\cr
\code{X, K, W} : Standard outputs of \code{fastICA()}. \code{X} is the pre-processed data matrix, \code{K} is the pre-whitening matrix projecting the data onto the first n principal components, and `W` is the estimated unmixing matrix.\cr
Three attributes are set within the list object. "ICAobject" for \code{class}, "ica" for \code{method} and "no" for \code{covar_cor}.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\examples{

data(expr_data)

ica_result <- runICA(expr_data)


}
