% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cluster_check.R
\name{sample_cluster_check}
\alias{sample_cluster_check}
\title{Function to estimate the number of clusters in IC coefficients or PC projections.}
\usage{
sample_cluster_check(input_list = NULL)
}
\arguments{
\item{input_list}{An input ICAobject or PCAobject.}
}
\value{
List with the following entries.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\keyword{keywords}

