% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Sample covariate information for 47 samples in \code{expr_data}}
\format{A matrix with 47 rows and 4 columns:
\describe{
  \item{gender}{Gender of the sample}
  \item{allergen}{Allergen that the sample was exposed to}
  \item{group}{Group of allergen}
  \item{level_of_reaction}{Level of reaction after allergen exposure}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60028}
}
\usage{
sample_info
}
\description{
A dataset containing the sample covariates for \code{expr_data}.
}
\details{
The original dataset is described in the following paper:
Dhingra N, Shemer A, Correa da Rosa J, Rozenblit M et al.
Molecular profiling of contact dermatitis skin identifies allergen-dependent differences in immune response.
J Allergy Clin Immunol 2014 Aug;134(2):362-72. PMID: 24768652
}
\keyword{datasets}
