% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_association_check.R
\name{covar_association_check}
\alias{covar_association_check}
\title{Calculate association between components and given covariates}
\usage{
covar_association_check(input_list = NULL, covars = NULL,
  cor_threshold = 0.05)
}
\arguments{
\item{input_list}{ICA or PCA object generated by \code{run_ica()} or
\code{run_pca()}.}

\item{covars}{A dataframe of covariates with each row for a sample
and each column for a covariate.}

\item{cor_threshold}{Bonferroni threshold that is going to be used for
identifying significant associations.}
}
\value{
The output will be the \code{input_list} with additional entries
        \code{comp_cov, covars, covar_threshold, covar_pvals}.
        \code{comp_cov} will contain a list with entries for each component
                        showing the associated covariate and p-value.
        \code{covars} will contain a copy of sample_info to keep all
                      the data in one place.
        \code{covar_threshold} will record the threshold
                               used to call significant associations.
        \code{covar_pvals} will contain the matrix of p-values
                           for all covariate and component pairs.
}
\description{
Check the association between sample covariates and
independent/principal components.
}
\examples{
data(expr_data, sample_info)

pca_object <- run_pca(expr_data)

pca_object <- covar_association_check(pca_object, covars = sample_info)
}
