% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clusters.R
\name{detect_clusters}
\alias{detect_clusters}
\title{Function to estimate the number of clusters in IC coefficients or PC projections.}
\usage{
detect_clusters(input_list = NULL)
}
\arguments{
\item{input_list}{input_list ICA or PCA object generated by \code{run_ica()} or \code{run_pca()}.}
}
\value{
Running \code{detect_clusters()} will generate a copy of the input_list with additional entries \code{n_clust,mclust_result}.
        \code{n_clust} will contain the number of estimated clusters for each component, and \code{mclust_result} is a list with
        each result of \code{mclust} saved for all components.
}
\description{
Performing ICA on a dataset and create a list object with results.
}
\examples{

data(expr_data)

pca_object <- run_pca(expr_data)
pca_object <- detect_clusters(pca_object)

}
