% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate_mx.R
\name{get_covariate_mx}
\alias{get_covariate_mx}
\title{Function that returns a matrix of IC coefficients that can be used as covariates.}
\usage{
get_covariate_mx(input_list = NULL, idx = NULL)
}
\arguments{
\item{input_list}{ICA or PCA object generated by \code{run_ica()} or \code{run_pca()}.}

\item{idx}{Component indexes that the user wishes to recover.}
}
\value{
Matrix with component coefficients
}
\description{
Returning a matrix of IC coefficients that are associated with covariates or have multiple clusters.
}
\examples{
data(expr_data)

ica_object <- run_ica(expr_data)
ica_object <- covar_association_check(ica_object, covars = sample_info)
ica_object <- detect_clusters(ica_object)
get_covariate_mx(ica_object)

}
\keyword{keywords}

