% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Plotting multiple ggplots}
\usage{
multiplot(plotlist = NULL, layout = NULL, cols = 1)
}
\arguments{
\item{plotlist}{list of plots}

\item{layout}{layout matrix for custom layout}

\item{cols}{number of columns}
}
\value{
single plot object
}
\description{
Equivalent function for ggplots to setting par(mfrow=c(x,y)) for
basic R plots.
Adapted from : http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\keyword{keywords}

