% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Custom PCA function for analyzing gene expression data.}
\usage{
run_pca(pheno_mx = NULL, scale_pheno = FALSE)
}
\arguments{
\item{pheno_mx}{Phenotype matrix with diemnsions g x N, where g is the number of genes and N is the number of samples.}

\item{scale_pheno}{Logical value specifying the scaling of row of the pheno_mx. Default is set to FALSE.}
}
\value{
The following entries will be generated in the output list \code{pca_object} after running the example above. \cr
\code{rotation} : Matrix of principal component gene weights where each column represents a single component. (standard \code{prcomp()} output)\cr
\code{x} : Matrix of the projections of the original data onto principal componets. Each column holds a projection. (standard \code{prcomp()} output)\cr
\code{sdev}: The standard deviation (square root of the eigen values) of each principal components. (standard \code{prcomp()} output)\cr
\code{percent_var} : The percent variance each principal component is explaining. Calculated based on \code{sdev}.\cr
\code{peaks} : Indicating which gene has a gene weight larger than 2 standard deviations of its component gene weights. \cr
\code{center} : The mean values for each gene used to center the data. (standard \code{prcomp()} output)\cr
\code{scale} : TRUE or FALSE value indicating whether the data was scaled. (standard \code{prcomp()} output)\cr
Three attributes are set within the list object. "PCAobject" for \code{class}, "pca" for \code{method} and "no" for \code{covar_cor}.\cr
}
\description{
Performing PCA on a dataset and create a list object with results.
}
\examples{

data(expr_data)

pca_object <- run_pca(expr_data)

}
